Download missing .sign files even if .cvd doesn't need to be updated
UX improvement for .sign file updates so it pulls missing .sign files
even if you don't need to download the associated cvd.
This may be meaningful for folks who have been running cvdupdate for a
while and already have main.cvd and bytecode.cvd, because we do not
publish those often.